Simulate plot errors in plant breeding trials

Creates a data frame with simulated plot errors for one or more traits in plant breeding trials across one or more environments. The plot errors consist of spatial error, random error and extraneous error. The spatial error is simulated according to either 1) bivariate interpolation using the interp function of the package 'interp', or 2) a separable first-order autoregressive process (AR1:AR1). The random error is simulated using an independent process. The extraneous error is simulated as the sum of column and/or row terms, where the user can choose from independent or correlated processes. The spatial, random and extraneous errors are combined according to a user-defined ratio.
For multiple traits, correlated errors can be simulated assuming 1) correlated spatial error between traits, 2) correlated random error between traits, 3) correlated extraneous error between traits, or 4) some combination of 1-3.
A separable correlation structure is assumed between traits and environments, but different variances can be assigned to different environment within trait combinations.

Usage

field_trial_error(
  n_envs = 1,
  n_traits = 1,
  n_reps = 2,
  n_cols = 10,
  n_rows = 20,
  rep_dir = "col",
  var_R = 1,
  S_cor_R = NULL,
  R_cor_R = NULL,
  E_cor_R = NULL,
  spatial_model = "Bivariate",
  complexity = NULL,
  plot_length = 10,
  plot_width = 2,
  col_cor = 0.4,
  row_cor = 0.6,
  prop_spatial = 0.5,
  prop_ext = 0,
  ext_dir = "row",
  ext_col_cor = 0,
  ext_row_cor = 0,
  return_effects = FALSE
)

Arguments

n_envs

Number of environments to be simulated (same as for compsym_asr_input or unstr_asr_output, where applicable). By default n_envs = 1.

n_traits

Number of traits to be simulated. By default, n_traits = 1.

n_reps

A vector specifying the number of complete replicates in each environment. If only one value is provided, all environments will be assigned the same number of replicates. By default, n_reps = 2.

n_cols

A vector specifying the total number of columns in each environment. If only one value is provided, all environments will be assigned the same number of columns. By default, n_cols = 10.

n_rows

A vector specifying the total number of rows in each environment. If only one value is provided, all environments will be assigned the same number of rows. By default n_rows = 20.

rep_dir

A vector specifying the direction of replicate blocks in each environment. Use "col" for a side-by-side arrangement (default), "row" for an above-and-below arrangement, or NA if only one replicate block is simulated. If only one value is provided, all environments will be assigned the same replicate block direction (where applicable).

var_R

A vector of error variances for each environment within trait combination. If only one value is provided, all environment within trait combinations will be assigned the same error variance.

S_cor_R

A matrix of spatial error correlations between more than one trait. If not specified, a diagonal matrix is constructed.

R_cor_R

A matrix of random error correlations between more than one trait. If not specified, a diagonal matrix is constructed.

E_cor_R

A matrix of extraneous error correlations between more than one trait. If not specified, a diagonal matrix is constructed. Note that the same correlation between traits is used for the column and row errors (where applicable).

spatial_model

A character string specifying the model used to simulate the two-dimensional spatial error term. One of either "Bivariate" (bivariate interpolation, the default) or "AR1:AR1" (separable first-order autoregressive process).

complexity

A vector specifying the complexity of the bivariate interpolation in each environment. If only one value is provided, all environments will be assigned the same complexity. If not specified and spatial_model = "Bivariate", the complexity is set to the maximum number of columns and rows in each environment. This generally provides good results. See interp for further details.

plot_length

A vector of plot lengths (column direction, usually longer side) for each environment. If only one value is provided, all environments will be assigned the same plot length. Only required when spatial_model = "Bivariate". By default, plot_length = 10.

plot_width

A vector of plot widths (row direction, usually shorter side) for each environment. If only one value is provided, all environments will be assigned the same plot width. Only required when spatial_model = "Bivariate". By default, plot_width = 2.

col_cor

A vector of column autocorrelations for each environment used in the AR1:AR1 spatial error model. If only one value is provided, all environments will be assigned the same column autocorrelation. Only required when spatial_model = "AR1:AR1". By default, col_cor = 0.4.

row_cor

A vector of row autocorrelations for each environment used in the AR1:AR1 spatial error model. If only one value is provided, all environments will be assigned the same row autocorrelation. Only required when spatial_model = "AR1:AR1". By default, row_cor = 0.6.

prop_spatial

A vector specifying the proportion of spatial error variance to total error variance (spatial + random + extraneous) for each environment within trait combination. If only one value is provided, all environment within trait combinations will be assigned the proportion of sptial error variance. By default, prop_spatial = 0.5.

prop_ext

A vector specifying the proportion of extraneous error variance to total error variance (spatial + random + extraneous) for each environment within trait combination. If only one value is provided, all environment within trait combinations will be assigned the proportion of extraneous error variance. By default, prop_ext = 0.

ext_dir

A vector specifying the direction of extraneous variation for each environment. Use "col" to simulate variation in the column direction, "row" (default) for variation in the row direction, "both" for variation in both directions, or NA if zero extraneous variation is simulated. When ext_dir = "both", half the variance is assigned to the columns and half is assigned to the rows. If only one value is provided, all environments will be assigned the same direction (where applicable).

ext_col_cor

A vector of column autocorrelations for each environment used in the extraneous error model. If only one value is provided, all environments will be assigned the same column autocorrelation. By default, ext_col_cor = 0.

ext_row_cor

A vector of row autocorrelations for each environment used in the extraneous error model. If only one value is provided, all environments will be assigned the same row autocorrelation. By default, ext_row_cor = 0.

return_effects

When TRUE, a list is returned with additional entries for each trait containing the spatial, random and extraneous errors. By default, return_effects = FALSE.

Value

A data frame with columns "env", "block", "col" and "row", as well as the simulated error for each trait. When return_effects = TRUE, a list is returned with additional columns for each trait providing the spatial, random and extraneous errors.

Examples

# Simulation of plot errors for two traits in three environments using a bivariate
# interpolation model for spatial variation.

n_envs <- 3
n_traits <- 2

# Field layout
n_cols <- 10
n_rows <- c(20, 30, 30)
rep_dir <- c("col", "row", "row")
plot_length <- 10
plot_width <- 2
n_reps <- c(2, 3, 3)

# Error variances for environment within trait combinations.
var_R <- c(0.2, 0.4, 0.6, 10, 15, 20)

# Spatial error correlations between traits 1 and 2.
S_cor_R <- matrix(
  c(
    1.0, 0.2,
    0.2, 1.0
  ),
  ncol = 2
)

# Structure of simulated error.
prop_spatial <- 0.4
prop_ext <- 0.2
ext_dir <- "row"
ext_row_cor <- -0.6

error_df <- field_trial_error(
  n_envs = n_envs,
  n_traits = n_traits,
  n_reps = n_reps,
  n_cols = n_cols,
  n_rows = n_rows,
  rep_dir = "row",
  var_R = var_R,
  S_cor_R = S_cor_R,
  spatial_model = "Bivariate",
  plot_length = plot_length,
  plot_width = plot_width,
  prop_spatial = prop_spatial,
  prop_ext = prop_ext,
  ext_dir = ext_dir,
  ext_row_cor = ext_row_cor,
  return_effects = TRUE
)